Move GPU base predictor to dedicated module
Having it in `__init__` needlessly requires all predictors in this import openAI module which is not necessary
