+++ custom_model_runner/datarobot_drum/drum/gpu_predictors/nemo_predictor.py
@@ -14,12 +14,12 @@ from requests import ConnectionError, Timeout
 
 from datarobot_drum.drum.enum import CustomHooks
 from datarobot_drum.drum.exceptions import DrumCommonException
-from datarobot_drum.drum.gpu_predictors import BaseGpuPredictor
+from datarobot_drum.drum.gpu_predictors.base import BaseOpenAiGpuPredictor
 from datarobot_drum.drum.gpu_predictors.utils import read_model_config
 from datarobot_drum.resource.drum_server_utils import DrumServerProcess
 
 
-class NemoPredictor(BaseGpuPredictor):
+class NemoPredictor(BaseOpenAiGpuPredictor):
     def __init__(self):
         super().__init__()
